some migration cleanup and moved the old demo_game.rb to examples/simple_demo
