--- db/migrate/20250323193327_create_players.rb
@@ -1,9 +1,0 @@
-class CreatePlayers < ActiveRecord::Migration[8.0]
-  def change
-    create_table :players do |t|
-      t.string :name
-
-      t.timestamps
-    end
-  end
-end
