--- db/migrate/20250401211627_add_state_to_game_sessions.rb
@@ -1,5 +1,0 @@
-class AddStateToGameSessions < ActiveRecord::Migration[8.0]
-  def change
-    add_column :game_sessions, :state, :jsonb
-  end
-end
