--- db/migrate/20250401211628_add_game_session_id_to_players.rb
@@ -1,5 +1,0 @@
-class AddGameSessionIdToPlayers < ActiveRecord::Migration[7.1]
-  def change
-    add_reference :players, :game_session, null: true, foreign_key: true
-  end
-end
