--- db/migrate/20250401211629_add_user_id_to_players.rb
@@ -1,5 +1,0 @@
-class AddUserIdToPlayers < ActiveRecord::Migration[7.1]
-  def change
-    add_reference :players, :user, null: true, foreign_key: true
-  end
-end
