--- db/migrate/20250401211630_drop_game_players.rb
@@ -1,5 +1,0 @@
-class DropGamePlayers < ActiveRecord::Migration[7.1]
-  def change
-    drop_table :game_players, if_exists: true
-  end
-end
