+++ db/migrate/20250402023344_add_devise_to_users.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+class AddDeviseToUsers < ActiveRecord::Migration[8.0]
+  def change
+    unless table_exists?(:users)
+      create_table :users do |t|
+        ## Database authenticatable
+        t.string :email,              null: false, default: ""
+        t.string :encrypted_password, null: false, default: ""
+
+        ## Recoverable
+        t.string   :reset_password_token
+        t.datetime :reset_password_sent_at
+
+        ## Rememberable
+        t.datetime :remember_created_at
+
+        t.timestamps null: false
+      end
+
+      add_index :users, :email,                unique: true
+      add_index :users, :reset_password_token, unique: true
+    end
+  end
+
+  def self.down
+    # By default, we don't want to make any assumption about how to roll back a migration when your
+    # model already existed. Please edit below which fields you would like to remove in this migration.
+    raise ActiveRecord::IrreversibleMigration
+  end
+end
