+++ examples/simple_demo/demo_game.rb
